requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.3
python-dotenv==1.0.1
